pyproject.toml
[project]
name = "proper-text-art"
version = "0.1.0"
description = "Converts pixel-art-style images into text-mode art (ANSI-colored Unicode or structural ASCII)."
authors = [{name = "Kenneth Allen"}]
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = [
    "text-art",
    "ascii-art",
    "ansi-art",
    "image-processing",
    "terminal",
    "generative-art",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Terminals",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
]

[project.urls]
Homepage = "https://github.com/caioluders/proper-text-art"
Repository = "https://github.com/caioluders/proper-text-art"
Issues = "https://github.com/caioluders/proper-text-art/issues"

[project.optional-dependencies]
web = ["gradio>=5.0.0"]
scripts = [
    "openai>=1.54.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
pta = "proper_text_art.cli:main"
pta-web = "proper_text_art.web:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.4.2",
    "ruff>=0.14.4",
]

[tool.hatch.build.targets.wheel]
packages = ["proper_text_art"]
exclude = ["assets"]

[tool.hatch.build.targets.sdist]
exclude = ["assets"]

[tool.ruff]
lint.select = ["I"]
